+++ examples/http/http.h
@@ -10,7 +10,9 @@
 typedef struct {
   int listen_fd;
   int epoll_fd;
-  cu_Vector clients; /**< stores client sockets */
+  cu_Vector clients;       /**< stores client sockets */
+  cu_SlabAllocator slab;   /**< slab allocator for buffers */
+  cu_Allocator slab_alloc; /**< allocator derived from slab */
 } cu_HttpServer;
 
 typedef enum {
@@ -21,7 +23,8 @@ typedef enum {
 
 CU_RESULT_DECL(cu_HttpServer, cu_HttpServer, cu_Http_Error)
 
-cu_HttpServer_Result cu_HttpServer_create(cu_Allocator allocator, uint16_t port);
+cu_HttpServer_Result cu_HttpServer_create(
+    cu_Allocator allocator, uint16_t port);
 void cu_HttpServer_destroy(cu_HttpServer *server);
 void cu_HttpServer_run(cu_HttpServer *server);