+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Changelog
+#Changelog
 
 All notable changes to this project will be documented in this file. See [conventional commits](https://www.conventionalcommits.org/) for commit guidelines.
 
@@ -89,6 +89,7 @@ All notable changes to this project will be documented in this file. See [conven
 
 - restructure build - ([12e1a30](https://git.schaub-dev.xyz/cppuniverse/libcute/commit/12e1a3039a1d0fbc68b6e6533f8b719b66df369a)) - Fabrice
 - add epoll based http server example
+- allocate request buffers from slab allocator for http example
 
 ### Fmt
 
@@ -153,4 +154,3 @@ All notable changes to this project will be documented in this file. See [conven
 - ensure aligned allocations in C allocator and use null-check macros
 
 - allow building tests with freestanding flags
-